+++ weebware - cheat/weebware/shared_funcs.h
@@ -3,7 +3,7 @@
 
 bool is_visible(c_base_entity* local, c_base_entity* target, int bone = 8);
 int get_config_index();
-long get_epoch_ms();
+long long get_epoch_ms();
 long get_epoch_s();
 float random_float(float flMinVal, float flMaxVal);
 void random_seed(UINT Seed);
